refactor(radio): extract useRttQueue custom state hook from queue component
