+++ assets/src/components/radio/rtts/index.ts
@@ -3,3 +3,4 @@ export * from "./queueItem"
 export * from "./detailsPanel"
 export * from "./activeBanner"
 export * from "./queue"
+export * from "./useRttQueue"
